+++ src/OneOf3.cs
@@ -9,7 +9,7 @@ namespace Grondo
     /// <typeparam name="T1">The first possible type.</typeparam>
     /// <typeparam name="T2">The second possible type.</typeparam>
     /// <typeparam name="T3">The third possible type.</typeparam>
-    [DebuggerDisplay("T{_index}({Value})")]
+    [DebuggerDisplay("T{Index}({Value})")]
     public readonly struct OneOf<T1, T2, T3> : IEquatable<OneOf<T1, T2, T3>>
     {
         private readonly T1? _value1;